+++ プロジェクト/Cliques_Engine/Project/Cliques_Engine/Cliques_Engine.vcxproj
@@ -182,6 +182,8 @@
     <ClInclude Include="C_Static_Model_User.h" />
     <ClInclude Include="C_Vertex_Index_System.h" />
     <ClInclude Include="C_Windows_Console_Log_System.h" />
+    <ClInclude Include="E_Blend_Mode.h" />
+    <ClInclude Include="E_Blend_Option.h" />
     <ClInclude Include="E_CONSOLE_LOG_COLOR.h" />
     <ClInclude Include="E_CONSOLE_LOG_TAGS.h" />
     <ClInclude Include="E_DEPTH_STENCIL_BUFFER_PIXEL_BYTE_FORMAT.h" />
@@ -190,6 +192,8 @@
     <ClInclude Include="E_MOUSE_CLICK_STATE.h" />
     <ClInclude Include="E_OS_KIND.h" />
     <ClInclude Include="E_RENDERING_API_KIND.h" />
+    <ClInclude Include="E_Rendering_Draw_Color.h" />
+    <ClInclude Include="E_Rendering_Draw_Format.h" />
     <ClInclude Include="E_RENDERING_SCREEN_FORMAT.h" />
     <ClInclude Include="E_SCREEN_TEXTURE_SETTING.h" />
     <ClInclude Include="GCC_Or_MSVC_Macro.h" />
